refactor(DocumentSigner): move SignatureForm and useSignEmployeeForm to Documents/shared
SignatureForm and useSignEmployeeForm are used by both DocumentSigner and
the standalone Employee.SignatureForm block, so they belong in the cross-
cutting Documents/shared/ layer rather than inside DocumentSigner/.

I9SignatureForm and EmploymentEligibility remain inside DocumentSigner/
so their API endpoints (i9_authorization, forms) are correctly attributed
to Employee.DocumentSigner in the endpoint inventory.
