+++ src/components/Employee/Documents/onboarding/DocumentSigner/I9SignatureForm/I9SignatureForm.tsx
@@ -1,7 +1,7 @@
 import { useTranslation, Trans } from 'react-i18next'
 import type { AuthorizationStatus } from '@gusto/embedded-api/models/components/i9authorization'
 import { useI9VerificationGetAuthorization } from '@gusto/embedded-api/react-query/i9VerificationGetAuthorization'
-import { useSignEmployeeForm, type PreparerFieldGroup } from '../shared/useSignEmployeeForm'
+import { useSignEmployeeForm, type PreparerFieldGroup } from '../../../shared/useSignEmployeeForm'
 import styles from './I9SignatureForm.module.scss'
 import {
   BaseComponent,
