+++ src/components/Employee/Documents/onboarding/DocumentSigner/documentSignerStateMachine.tsx
@@ -1,8 +1,8 @@
 import { type Form } from '@gusto/embedded-api/models/components/form'
-import { DocumentList } from './DocumentList/DocumentList'
-import { SignatureForm } from './SignatureForm/SignatureForm'
+import { SignatureForm } from '../../shared/SignatureForm/SignatureForm'
 import { EmploymentEligibility } from './EmploymentEligibility'
 import { I9SignatureForm } from './I9SignatureForm/I9SignatureForm'
+import { DocumentList } from './DocumentList/DocumentList'
 import { useFlow, type FlowContextInterface } from '@/components/Flow/useFlow'
 import type { componentEvents } from '@/shared/constants'
 import { ensureRequired } from '@/helpers/ensureRequired'
